Refactor Go plugin
