+++ test/plugins/go.test.js
@@ -1,12 +1,12 @@
 import assert from 'assert';
-import goUniversal from '../../lib/resolver/go-universal.js';
+import goUniversal from '../../lib/plugins/go';
 
 describe('go-universal', () => {
   const path = 'octo/dog.go';
 
   it('resolves local file', () => {
     assert.deepEqual(
-      goUniversal({ path, target: '../foo' }),
+      goUniversal.resolve({ path, target: '../foo' }),
       [
         '{BASE_URL}foo/foo.go',
         '{BASE_URL}foo.go',
@@ -17,7 +17,7 @@ describe('go-universal', () => {
 
   it('resolves package', () => {
     assert.deepEqual(
-      goUniversal({ path, target: 'foo' }),
+      goUniversal.resolve({ path, target: 'foo' }),
       [
         'https://foo',
         'https://golang.org/pkg/foo',
@@ -27,7 +27,7 @@ describe('go-universal', () => {
 
   it('resolves github shorthand', () => {
     assert.deepEqual(
-      goUniversal({ path, target: 'github.com/foo/bar' }),
+      goUniversal.resolve({ path, target: 'github.com/foo/bar' }),
       [
         '{BASE_URL}/foo/bar/blob/master/bar.go',
         '{BASE_URL}/foo/bar',
@@ -37,7 +37,7 @@ describe('go-universal', () => {
 
   it('resolves github deep link', () => {
     assert.deepEqual(
-      goUniversal({ path, target: 'github.com/foo/bar/baze' }),
+      goUniversal.resolve({ path, target: 'github.com/foo/bar/baze' }),
       [
         '{BASE_URL}/foo/bar/blob/master/baze/baze.go',
         '{BASE_URL}/foo/bar/tree/master/baze',
